@@ -7,31 +7,31 @@
   },
   "devDependencies": {
     "@tsconfig/node16": "1.0.3",
-    "@types/jest": "28.1.4",
+    "@types/jest": "28.1.6",
     "@types/mkdirp": "1.0.2",
     "@types/node": "ts3.7",
     "@types/rimraf": "3.0.2",
     "@types/pg": "8.6.5",
-    "@typescript-eslint/eslint-plugin": "5.30.4",
-    "@typescript-eslint/parser": "5.30.4",
-    "eslint": "8.19.0",
+    "@typescript-eslint/eslint-plugin": "5.30.7",
+    "@typescript-eslint/parser": "5.30.7",
+    "eslint": "8.20.0",
     "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-jest": "26.5.3",
+    "eslint-plugin-jest": "26.6.0",
     "eslint-plugin-prettier": "4.2.1",
-    "jest": "28.1.2",
-    "jest-environment-node": "28.1.2",
+    "jest": "28.1.3",
+    "jest-environment-node": "28.1.3",
     "jest-junit": "14.0.0",
     "mkdirp": "1.0.4",
     "prettier": "2.7.1",
     "rimraf": "3.0.2",
-    "testcontainers": "8.11.0",
-    "ts-jest": "28.0.5",
-    "ts-node": "10.8.2",
+    "testcontainers": "8.12.0",
+    "ts-jest": "28.0.7",
+    "ts-node": "10.9.1",
     "typescript": "4.7.4"
   },
   "scripts": {
     "lint": "eslint --ignore-path .gitignore .",
-    "test": "jest --runInBand",
-    "test:ci": "jest --runInBand --config=jest.ci.config.ts"
+    "test": "jest",
+    "test:ci": "jest --config=jest.ci.config.ts"
   }
 }
