@@ -4,30 +4,26 @@ import hu.bsstudio.bssweb.event.model.DetailedEvent
 import hu.bsstudio.bssweb.event.service.EventService
 import hu.bsstudio.bssweb.eventvideo.entity.EventVideoEntity
 import hu.bsstudio.bssweb.eventvideo.repository.EventVideoRepository
-import io.mockk.MockKAnnotations
 import io.mockk.every
+import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
+import io.mockk.junit5.MockKExtension
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import java.time.LocalDate
 import java.util.Optional
 import java.util.UUID
 
+@ExtendWith(MockKExtension::class)
 internal class DefaultEventVideoServiceTest {
 
     @MockK
     private lateinit var mockRepository: EventVideoRepository
     @MockK
     private lateinit var mockEventService: EventService
-
+    @InjectMockKs
     private lateinit var underTest: DefaultEventVideoService
-
-    @BeforeEach
-    fun setUp() {
-        MockKAnnotations.init(this)
-        underTest = DefaultEventVideoService(mockRepository, mockEventService)
-    }
 
     @Test
     fun addVideoToEvent() {
