@@ -3,26 +3,23 @@ package hu.bsstudio.bssweb.videocrew.controller
 import hu.bsstudio.bssweb.videocrew.model.SimpleCrew
 import hu.bsstudio.bssweb.videocrew.model.VideoCrew
 import hu.bsstudio.bssweb.videocrew.service.VideoCrewService
-import io.mockk.MockKAnnotations
 import io.mockk.every
+import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
+import io.mockk.junit5.MockKExtension
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.http.HttpStatus
 import java.util.UUID
 
+@ExtendWith(MockKExtension::class)
 internal class VideoCrewControllerTest {
 
     @MockK
     private lateinit var mockService: VideoCrewService
+    @InjectMockKs
     private lateinit var underTest: VideoCrewController
-
-    @BeforeEach
-    internal fun setUp() {
-        MockKAnnotations.init(this)
-        this.underTest = VideoCrewController(mockService)
-    }
 
     @Test
     internal fun addPosition() {
@@ -30,7 +27,7 @@ internal class VideoCrewControllerTest {
 
         val response = this.underTest.addPosition(VIDEO_ID, POSITION, MEMBER_ID)
 
-        assertThat(response.statusCode).isEqualTo(HttpStatus.OK)
+        assertThat(response.statusCode).isEqualTo(HttpStatus.CREATED)
         assertThat(response.body).isEqualTo(VIDEO_CREW_LIST)
     }
 
